# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Тех. Задание Ozon",
        "title": "Ozon Backend Test",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/link": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "link"
                ],
                "summary": "Create Token",
                "operationId": "create-token",
                "parameters": [
                    {
                        "description": "link",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "token",
                        "schema": {
                            "$ref": "#/definitions/v1.tokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/link/{token}": {
            "get": {
                "tags": [
                    "link"
                ],
                "summary": "Get long URL",
                "operationId": "get-long-url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "link",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "link",
                        "schema": {
                            "$ref": "#/definitions/v1.linkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/v1.errResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreateTokenRequest": {
            "type": "object",
            "properties": {
                "original_url": {
                    "type": "string"
                }
            }
        },
        "v1.errResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "v1.linkResponse": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                }
            }
        },
        "v1.tokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}